+++ file-transfer/db/mongo.go
@@ -1,4 +1,4 @@
-package main
+package db
 
 import (
 	"context"
